+++ internal/manager/config_validation.go
@@ -10,6 +10,7 @@ import (
 	k8stypes "k8s.io/apimachinery/pkg/types"
 
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/adminapi"
+	"github.com/kong/kubernetes-ingress-controller/v3/internal/clients"
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/konnect"
 	cfgtypes "github.com/kong/kubernetes-ingress-controller/v3/internal/manager/config/types"
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/manager/featuregates"
@@ -75,6 +76,9 @@ func (c *Config) Validate() error {
 	if err := c.validateFallbackConfiguration(); err != nil {
 		return fmt.Errorf("invalid fallback config settings: %w", err)
 	}
+	if err := c.validateGatewayDiscovery(); err != nil {
+		return fmt.Errorf("invalid gateway discovery configuration: %w", err)
+	}
 
 	return nil
 }
@@ -118,6 +122,22 @@ func (c *Config) validateFallbackConfiguration() error {
 			"--use-last-valid-config-for-fallback or CONTROLLER_USE_LAST_VALID_CONFIG_FOR_FALLBACK can only be used with %s feature gate enabled",
 			featuregates.FallbackConfiguration,
 		)
+	}
+	return nil
+}
+
+func (c *Config) validateGatewayDiscovery() error {
+	// Skip validation if gateway discovery is not enabled.
+	if _, ok := c.KongAdminSvc.Get(); !ok {
+		return nil
+	}
+
+	if c.GatewayDiscoveryReadinessCheckInterval < clients.MinReadinessReconciliationInterval {
+		return fmt.Errorf("Readiness check reconciliation interval cannot be less than %s",
+			clients.MinReadinessReconciliationInterval)
+	}
+	if c.GatewayDiscoveryReadinessCheckTimeout >= c.GatewayDiscoveryReadinessCheckInterval {
+		return fmt.Errorf("Readiness check timeout must be less than readiness check recociliation interval")
 	}
 	return nil
 }